+++ src/Common/NetworkCore/Network/Pipeline/Inbound/ValidationInboundPipelineStage.cs
@@ -33,7 +33,8 @@ public class ValidationInboundPipelineStage : IInboundPipelineStage
     public ValidationInboundPipelineStage(NullEventDetailsValidator nullValidator,
         EntityDefinitionEventDetailsValidator entityDefinitionValidator,
         WorldSegmentSubscriptionEventDetailsValidator worldSegmentSubscriptionValidator,
-        MoveEventDetailsValidator moveValidator, RequestMoveEventDetailsValidator requestMoveValidator)
+        MoveEventDetailsValidator moveValidator, RequestMoveEventDetailsValidator requestMoveValidator,
+        EntityGridPositionEventDetailsValidator entityGridPositionValidator)
     {
         this.nullValidator = nullValidator;
         validators = new Dictionary<EventId, IEventDetailsValidator>
@@ -44,7 +45,8 @@ public class ValidationInboundPipelineStage : IInboundPipelineStage
             { EventId.Core_WorldManagement_Unsubscribe, worldSegmentSubscriptionValidator },
             { EventId.Client_EntitySynchronization_Update, entityDefinitionValidator },
             { EventId.Core_Movement_Move, moveValidator },
-            { EventId.Core_Movement_RequestMove, requestMoveValidator }
+            { EventId.Core_Movement_RequestMove, requestMoveValidator },
+            { EventId.Core_WorldManagement_EntityLeaveWorldSegment, entityGridPositionValidator }
         };
     }