+++ src/Server/ServerCore/Systems/WorldManagement/WorldManagementInstaller.cs
@@ -57,5 +57,8 @@ public sealed class WorldManagementInstaller : IWindsorInstaller
 
         container.Register(Component.For<WorldManagementServices>()
             .LifestyleSingleton());
+
+        container.Register(Component.For<WorldSegmentChangeMonitor>()
+            .LifestyleSingleton());
     }
 }